Rename dynamic-image to dynamic-image-base. Improve obtaining uri image in BaseDynamicImage and it's implementations. Fix project url. Improve DynamicImage class docs.
