+++ dynamic/image-base/.gitignore
