+++ dynamic/image-base/src/main/java/com/twocoders/dynamic/image/base/DynamicImageBindingAdapter.kt
@@ -1,8 +1,8 @@
-package com.twocoders.dynamic.image.coil
+package com.twocoders.dynamic.image.base
 
 import android.widget.ImageView
 import androidx.databinding.BindingAdapter
 
 @BindingAdapter(value = ["android:src", "loadWithCrossFade"], requireAll = false)
-fun ImageView.loadDynamicImage(image: DynamicImage, withCrossFade: Boolean = false) =
+fun ImageView.loadDynamicImage(image: BaseDynamicImage, withCrossFade: Boolean = false) =
     image.loadDrawableInto(this, withCrossFade)
