--- dynamic/image-glide/src/main/java/com/twocoders/dynamic/image/glide/DynamicImageBindingAdapter.kt
@@ -1,8 +1,0 @@
-package com.twocoders.dynamic.image.glide
-
-import android.widget.ImageView
-import androidx.databinding.BindingAdapter
-
-@BindingAdapter(value = ["android:src", "loadWithCrossFade"], requireAll = false)
-fun ImageView.loadDynamicImage(image: DynamicImage, withCrossFade: Boolean = false) =
-    image.loadDrawableInto(this, withCrossFade)
