+++ dynamic/image-picasso/src/main/java/com/twocoders/dynamic/image/picasso/DynamicImage.kt
@@ -11,19 +11,20 @@ import androidx.annotation.ColorInt
 import androidx.annotation.DrawableRes
 import com.squareup.picasso.Picasso
 import com.squareup.picasso.Target
-import com.twocoders.dynamic.image.BaseDynamicImage
-import com.twocoders.dynamic.image.component.UriComponent
+import com.twocoders.dynamic.image.base.BaseDynamicImage
+import com.twocoders.dynamic.image.base.component.UriComponent
 
 /**
  *
- * Handy class which can be used to bind image data to views using them.
+ * Handy class which can be used to bind image data to views.
  * Data can be in [DrawableRes], [Drawable], or [UriComponent] format.
  *
  * Use one of [DynamicImage.from] creator methods to create the data
  * and [DynamicImage.getDrawable] to obtain the final [Drawable] output. Or you can
  * provide target [ImageView] class into the [DynamicImage.loadDrawableInto] method.
  *
- * This [DynamicImage] implementation uses [Picasso], for other implementations please visit our GitHub.
+ * This [DynamicImage] implementation uses [Picasso], for other implementations please visit our GitHub:
+ * [Android Dynamic Components](https://github.com/Two-Coders/android-dynamic-components)
  *
  */
 @Suppress("unused", "MemberVisibilityCanBePrivate")
@@ -57,33 +58,30 @@ open class DynamicImage : BaseDynamicImage {
         }
     }
 
-    override suspend fun getDrawable(context: Context): Drawable? {
-        imageUri?.let { uri ->
-            return BitmapDrawable(
-                context.resources,
-                Picasso.get()
-                    .load(uri.image)
-                    .get()
-            )
-        }
+    override suspend fun getDrawableFromUri(
+        context: Context,
+        imageUriComponent: UriComponent
+    ) = BitmapDrawable(
+        context.resources,
+        Picasso.get()
+            .load(imageUriComponent.image)
+            .get()
+    )
 
-        return super.getDrawable(context)
-    }
-
-    override fun getDrawable(context: Context, callback: (drawable: Drawable) -> Unit) {
-        imageUri?.let { imageUriComponent ->
-            Picasso.get()
-                .load(imageUriComponent.image)
-                .into(object : Target {
-                    override fun onBitmapLoaded(bitmap: Bitmap?, from: Picasso.LoadedFrom?) {
-                        callback(BitmapDrawable(context.resources, bitmap))
-                    }
-                    override fun onBitmapFailed(e: Exception?, errorDrawable: Drawable?) {}
-                    override fun onPrepareLoad(placeHolderDrawable: Drawable?) {}
-                })
-        }
-
-        super.getDrawable(context, callback)
+    override fun getDrawableFromUri(
+        context: Context,
+        imageUriComponent: UriComponent,
+        callback: (drawable: Drawable) -> Unit
+    ) {
+        Picasso.get()
+            .load(imageUriComponent.image)
+            .into(object : Target {
+                override fun onBitmapLoaded(bitmap: Bitmap?, from: Picasso.LoadedFrom?) {
+                    callback(BitmapDrawable(context.resources, bitmap))
+                }
+                override fun onBitmapFailed(e: Exception?, errorDrawable: Drawable?) {}
+                override fun onPrepareLoad(placeHolderDrawable: Drawable?) {}
+            })
     }
 
     override fun loadDrawableInto(
