--- dynamic/image/src/main/AndroidManifest.xml
@@ -1,1 +1,0 @@
-<manifest package="com.twocoders.dynamic.image" />
