+++ settings.gradle
@@ -1,6 +1,6 @@
 include ':dynamic:text',
         ':dynamic:color',
-        ':dynamic:image',
+        ':dynamic:image-base',
         ':dynamic:image-coil',
         ':dynamic:image-glide',
         ':dynamic:image-picasso'
